services/policy-engine/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
pydantic==2.6.1
pydantic-settings==2.1.0
redis[hiredis]==5.0.1
nats-py==2.7.2
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.25.3
structlog==24.1.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
prometheus-client==0.20.0
pytest==8.0.1
pytest-asyncio==0.23.5
